telegram-bot-aiogram/requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.0
aiomysql>=0.2.0
aioredis>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
python-magic>=0.4.0
pytest>=7.4.0
pytest-asyncio>=0.23.0